move getMintPubkeyFromTokenAccountPubkey into token index file
